requirements.txt
flask==2.0.1
flask-cors==3.1.1
python-dotenv==0.19.1